requirements.txt
torchvision
joblib
tqdm
path.py
pillow
msgpack
spatial-correlation-sampler
scipy
blessings
progressbar2
tensorboardX
pypng
scikit-image
opencv-python
